requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
pyasn1==0.5.1
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
urllib3==2.1.0
yarl==1.9.3
